requirements.txt
fastapi
uvicorn[standard]
python-multipart
pillow
numpy
tensorflow-cpu==2.15.0
gdown